Move UnitList.vue file
